requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.27
python-dotenv==1.0.1
openai==1.12.0
fake-useragent==1.4.0
httpx==0.27.0
tabulate>=0.9.0
pandas>=2.2.0
tiktoken>=0.5.0 